chore: add circle ci (#20)
adds jobs for running tests and for building in docker. We can make this stricter by adding go vet and friends if wanted. The next step with this is to get the tests running with the docker compose config, but I want to land the simple version first.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>