chore: push docker tags from ci (#26)
Adds script and ci config to push to docker hub when
- the master or main branch changes
- semver like tags are pushed to the repo

Pushing a semver like tag pushes a docker image with the same tag. The docker latest tag points to the last semver version tag

Pushing changes to main or master gets you a docker tag of the form <branch name>-<date>-<truncated commit sha>
e.g main-2020-f435621. a <branch>-latest docker tag is provided as a convenince if you really want the bleeding edge. The branch name is included in the docker tag name to make that clearer.

This is a tweak on what I set up for go-ipfs in https://github.com/ipfs/go-ipfs/pull/6949

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>