feat: refactor task scheduler implementation
This is a major refactoring of the scheduling system for Visor tasks. This is the additive change, a subsequent
commit will remove the old scheduler code including redis.

A new Scheduler type is added that runs a set of Tasks, restarting them when they fail. Concurrency levels can
be managed by adding more or less tasks of a specific type. Each task runs in a single goroutine, usually in
a loop to process work in batches.

New tasks added (in tasks package):

 - ChainHeadIndexer - follows the chain head recording new tipsets
 - ChainHistoryIndexer - walks the chain history recording historic tipsets
 - ActorStateChangeProcessor - processes tipsets to extract actor state changes
 - ActorStateProcessor - processes actor state changes found by the ActorStateChangeProcessor
 - MessageProcessor - processes tipsets to extract messages

The postgres database is used as a global coordination point. ChainHeadIndexer and ChainHistoryIndexer write
newly discovered tipsets to the visor_processing_statechanges and visor_processing_messages tables. The
ActorStateChangeProcessor and MessageProcessor tasks lease tipsets from these tables in a batch to process.
The ActorStateChangeProcessor writes actor information to the visor_processing_actors table which
the ActorStateProcessor task uses to determine which actors require processing.

Tasks may be marked as global singletons by using a GlobalLock when registering with the Scheduler. The
GlobalLock takes an exclusive advisory lock in postgres for the duration of the session which ensures that
only one task of that type may run anywhere. The ChainHeadIndexer and ChainHistoryIndexer tasks are expected
to be global singletons.

A new cli subcommand called run is added which instantiates the new scheduler with a sample number of each type
of task. This is suitable for testing but further configuration of task arrangement and concurrency needs to
be added in the future. The index and process subcommands will be changed to use the new scheduler in a later
change.