feat: add schema migration capability (#40)
* feat: add schema migration capability

Introduce schema versioning and migrations between versions. The database schema is versioned using
a simple monotonic counter.

Migrations are placed in storage/migrations and define the steps needed to migrate to and from
different database schema versions.

A new `migrate` subcommand is added. When run without arguments it checks the version of the database schema
in the connected database and verifies that it is compatible with the current version required by visor.

The `--latest` flag instructs the migrate command to migrate the database schema to the latest version.

The `--to` flag instructs the migrate command to migrate to a specific version, if possible.

By default the `index` and `process` subcommands verify the schema version but do not automatically migrate to
the latest. This behaviour can be enabled by supplying the `--allow-schema-migrations` flag.

Every running instance of visor now takes a shared advisory lock in the database to indicate their presence.
A schema migration requires an exclusive advisory lock which will fail if there are any existing locks already
taken. This ensures that migrations are only performed by a single instance of visor.

* Review changes