feat: add deferred migrations
Deferred migrations are migrations that do not need to be applied immediately and may be
slow to execute. An example would be adding a new index to an existing large table.
A normal migration would lock the entire database until the indexing has completed which
may be unacceptably long.

All pending deferred migrations may be applied at a later date using `visor migrate --deferred`
or a up a specific schema version by using `visor migrate --deferred-to=VERSION`

Schema migration 9 is modified to use deferred migrations for the new indexes.