feat: rpc reconnection on failure (#149)
* feat: rpc reconnection on failure

Tasks are now passed an APIOpener which they use to acquire a lens. On
encountering a fatal error the task exits and closes the lens. The scheduler
restarts the task after a delay allowing the task to attempt a new connection
to the lens.

The lotus lens opener returns a new rpc connection. The lotusrepo and carrepo
openers return a shared instance of their repo since they hold exclusive locks
over the store.

* Don't return context when creating lens openers

* Remove lotus lens opener dependency on cached cli context