feat: allow actor state processor to run without leasing (#178)
* feat: allow actor state processor to run without leasing

Leasing of actor states to process causes high contention on the visor_processing_actors
table since many instances are competing to lock rows to update their leases.

With this change, when --actorchange-lease is set to zero then the actor change processor
will not use leasing at all. Instead it assumes that it can work on any actor states
between the height ranges specified with --from and --to. It further subdivides that
height range by the number of workers started so each worker gets its own range of
heights to process.

Note that this requires manual coordination of height ranges allocated to each
instance of visor otherwise each instance may duplicate work performed by another
instance.

Note also that --to should be set to a reasonable value, typically under the current
height of the chain. Since the range specified with --to and --from is subdivided
evenly between workers a very large upper bound could result in some workers being
allocated ranges that have no data on the chain.

* fix: Remove magic numbers

Co-authored-by: Mike Greenberg <mg+github@protocol.ai>