feat(lens): Optimize StateGetActor calls.
Add a utility function called "OptimizedStateGetActorWithFallback()" which
gets the current tipset's state by trying to load the child's state and using
ParentState() on that, so that we avoid recomputing the state for the current
tipset by applying messages.

If anything goes wrong, the function falls back to the original, more
expensive StateGetActor, and logs a warning.