feat(gas outputs): Add Height and ActorName (#270)
* Feat(gas outputs): Add Height and ActorName

Gas outputs are very important for many base-fee and gas-usage related views.

The table has no height, which means it needs to be joined with blocks all the
time on something like state_root to get a timestamp. This is likely
innefficient.

Additionally, it needs to be joined with actors to get the actor code. Since
we are "deriving" and already include redundant info, we might also get this
one. This allows easy filtering of messages by actor type and method.

* feat(gas outputs): Migrations: add height and actor_name columns

* Migration 22: set height to bigint

* feat(migrations): add constrains to new derived_gas_outputs columns

* feat(migrations): add height column as primary key in derived_gas_outputs.

* feat(migrations): re-index gasoutputs_height_code migration to 24.

* feat(migrations): fix gas_output_migration.

* Add state_root to derived_gas_outputs index

Co-authored-by: Ian Davis <nospam@iandavis.com>