update sql lens to archival interface
This change updates the SqlRepo Lens to provide a datastore backing using the
same code as in
https://github.com/filecoin-project/statediff/pull/119
and
https://github.com/ribasushi/ltsh/tree/psql_fullcapture

ref
https://github.com/ribasushi/ltsh/commit/4654e05d922

The changed interface uses a separate table for tracking 'current', and headers.

Possible remaining TODO is to improve the query used for getting the
current tipset to respect the lookback parameter and instead find the
tipset at height `n` before the most recently known tipset.