chore: add support for v5 actors (#554)
Update to actors v5-rc3. The actor methods and states should not change between rc and release so there should not be any difficulty on final upgrade.

Update to latest statediff which I have updated to v5 too (see here). We only use statediff for generating message parameter type descriptions.

We should be good to parse the new miner messages and the signature change for ComputeDataCommitment.

I reworked a lot of places where we were importing codes from each actors version to use our generated actor shims and added some helper methods to the templates.

(Ideally I would like to merge the statediff message parameter generation and the actor shim generation into a single package but this is not the best time to do that.)

- also fix #549 & #547

Co-authored-by: frrist <forrest@protocol.ai>